5.0.5.EL

+++ drivers/char/n_tty.c
@@ -1157,7 +1157,7 @@ static inline int copy_from_read_buf(struct tty_struct *tty,
 	retval = 0;
 	spin_lock_irqsave(&tty->read_lock, flags);
 	n = min(tty->read_cnt, N_TTY_BUF_SIZE - tty->read_tail);
-	n = min((ssize_t)*nr, n);
+	n = min(*nr, n);
 	spin_unlock_irqrestore(&tty->read_lock, flags);
 	if (n) {
 		mb();
